Module pallet_liquid_staking::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageExchangeRate

_GeneratedPrefixForStorageInsurancePool

_GeneratedPrefixForStorageLiquidCurrency

_GeneratedPrefixForStorageMatchingPool

_GeneratedPrefixForStorageReserveFactor

_GeneratedPrefixForStorageStakingCurrency

_GeneratedPrefixForStorageStakingPool

_GeneratedPrefixForStorageStakingPoolCapacity

_GeneratedPrefixForStorageTotalSlashed

_GeneratedPrefixForStorageUnstakeQueue

_GeneratedPrefixForStorageXcmFeesCompensation

_GeneratedPrefixForStorageXcmWeight

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

AssetIdOf

BalanceOf

ExchangeRate

The exchange rate between relaychain native asset and the voucher.

InsurancePool

Total amount of charged assets to be used as xcm fees.

LiquidCurrency

Liquid currency asset id

MatchingPool

Store total stake amount and unstake amount in each era, And will update when stake/unstake occurred.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ReserveFactor

Fraction of reward currently set aside for reserves.

StakingCurrency

Staking currency asset id

StakingPool

Total amount of staked assets on relaycahin.

StakingPoolCapacity

Staking pool capacity

TotalSlashed

Total amount of slash in relaychain

UnstakeQueue

Manage which we should pay off to.

XcmFeesCompensation

Relaychain xcm fees compensation

XcmWeight

Xcm weight in BuyExecution