Trait pallet_liquid_staking::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Currency: MultiCurrencyExtended<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance, Amount = Amount>;
    type StakingCurrency: Get<CurrencyId>;
    type LiquidCurrency: Get<CurrencyId>;
    type PalletId: Get<PalletId>;
    type WithdrawOrigin: EnsureOrigin<Self::Origin>;
    type MaxWithdrawAmount: Get<Balance>;
    type MaxAccountProcessingUnstake: Get<u32>;
    type WeightInfo: WeightInfo;
    type XcmTransfer: XcmTransfer<Self::AccountId, Balance, CurrencyId>;
    type Members: SortedMembers<Self::AccountId>;
    type BaseXcmWeight: Get<Weight>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

The overarching event type.

type Currency: MultiCurrencyExtended<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance, Amount = Amount>

Currency type used for staking and liquid assets

type StakingCurrency: Get<CurrencyId>

Currency used for staking

type LiquidCurrency: Get<CurrencyId>

Currency used for liquid voucher

type PalletId: Get<PalletId>

The pallet id of liquid staking, keeps all the staking assets.

type WithdrawOrigin: EnsureOrigin<Self::Origin>

The origin which can withdraw staking assets.

type MaxWithdrawAmount: Get<Balance>

The maximum assets can be withdrawed to a multisig account.

type MaxAccountProcessingUnstake: Get<u32>

The maximum size of AccountProcessingUnstake

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type XcmTransfer: XcmTransfer<Self::AccountId, Balance, CurrencyId>

XCM transfer

type Members: SortedMembers<Self::AccountId>

Approved agent list on relaychain

type BaseXcmWeight: Get<Weight>

Base xcm weight to use for cross chain transfer

Loading content...