Trait pallet_liquid_staking::pallet::Config

pub trait Config: Config + Config + Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Origin: IsType<<Self as Config>::Origin> + Into<Result<Origin, <Self as Config>::Origin>>;
    type Call: IsType<<Self as Config>::Call> + From<Call<Self>>;
    type Assets: Transfer<Self::AccountId, AssetId = CurrencyId> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + InspectMetadata<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
    type RelayOrigin: EnsureOrigin<<Self as Config>::Origin>;
    type UpdateOrigin: EnsureOrigin<<Self as Config>::Origin>;
    type Members: SortedMembers<Self::AccountId>;
    type PalletId: Get<PalletId>;
    type SelfParaId: Get<ParaId>;
    type DerivativeIndexList: Get<Vec<DerivativeIndex>>;
    type XcmFees: Get<BalanceOf<Self>>;
    type StakingCurrency: Get<AssetIdOf<Self>>;
    type LiquidCurrency: Get<AssetIdOf<Self>>;
    type MinStake: Get<BalanceOf<Self>>;
    type MinUnstake: Get<BalanceOf<Self>>;
    type WeightInfo: WeightInfo;
    type BondingDuration: Get<EraIndex>;
    type MinNominatorBond: Get<BalanceOf<Self>>;
    type EraLength: Get<BlockNumberFor<Self>>;
    type NumSlashingSpans: Get<u32>;
    type RelayChainValidationDataProvider: ValidationDataProvider + BlockNumberProvider<BlockNumber = BlockNumberFor<Self>>;
    type XCM: XcmHelper<Self, BalanceOf<Self>, AssetIdOf<Self>, Self::AccountId>;
    type DistributionStrategy: DistributionStrategy<BalanceOf<Self>>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Origin: IsType<<Self as Config>::Origin> + Into<Result<Origin, <Self as Config>::Origin>>

type Call: IsType<<Self as Config>::Call> + From<Call<Self>>

type Assets: Transfer<Self::AccountId, AssetId = CurrencyId> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + InspectMetadata<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Assets for deposit/withdraw assets to/from pallet account

type RelayOrigin: EnsureOrigin<<Self as Config>::Origin>

The origin which can do operation on relaychain using parachain’s sovereign account

type UpdateOrigin: EnsureOrigin<<Self as Config>::Origin>

The origin which can update liquid currency, staking currency and other parameters

type Members: SortedMembers<Self::AccountId>

Approved accouts which can call withdraw_unbonded and settlement

type PalletId: Get<PalletId>

The pallet id of liquid staking, keeps all the staking assets

type SelfParaId: Get<ParaId>

Returns the parachain ID we are running with.

type DerivativeIndexList: Get<Vec<DerivativeIndex>>

Derivative index list

type XcmFees: Get<BalanceOf<Self>>

Xcm fees

type StakingCurrency: Get<AssetIdOf<Self>>

Staking currency

type LiquidCurrency: Get<AssetIdOf<Self>>

Liquid currency

type MinStake: Get<BalanceOf<Self>>

Minimum stake amount

type MinUnstake: Get<BalanceOf<Self>>

Minimum unstake amount

type WeightInfo: WeightInfo

Weight information

type BondingDuration: Get<EraIndex>

Number of unbond indexes for unlocking.

type MinNominatorBond: Get<BalanceOf<Self>>

The minimum active bond to become and maintain the role of a nominator.

type EraLength: Get<BlockNumberFor<Self>>

Number of blocknumbers that each period contains. SessionsPerEra * EpochDuration / MILLISECS_PER_BLOCK

type NumSlashingSpans: Get<u32>

type RelayChainValidationDataProvider: ValidationDataProvider + BlockNumberProvider<BlockNumber = BlockNumberFor<Self>>

The relay’s validation data provider

type XCM: XcmHelper<Self, BalanceOf<Self>, AssetIdOf<Self>, Self::AccountId>

To expose XCM helper functions

type DistributionStrategy: DistributionStrategy<BalanceOf<Self>>

Currenty strategy for distributing assets to multi-accounts