Trait pallet_liquid_staking::pallet::Config

pub trait Config: Config + Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Assets: Transfer<Self::AccountId, AssetId = CurrencyId> + Mutate<Self::AccountId, Balance = Balance>;
    type RelayOrigin: EnsureOrigin<Self::Origin>;
    type UpdateOrigin: EnsureOrigin<Self::Origin>;
    type PalletId: Get<PalletId>;
    type XcmSender: SendXcm;
    type SelfParaId: Get<ParaId>;
    type DerivativeIndex: Get<u16>;
    type AccountIdToMultiLocation: Convert<Self::AccountId, MultiLocation>;
    type UnstakeQueueCapacity: Get<u32>;
    type MaxRewardsPerEra: Get<BalanceOf<Self>>;
    type MaxSlashesPerEra: Get<BalanceOf<Self>>;
    type MinStakeAmount: Get<BalanceOf<Self>>;
    type MinUnstakeAmount: Get<BalanceOf<Self>>;
    type RelayNetwork: Get<NetworkId>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Assets: Transfer<Self::AccountId, AssetId = CurrencyId> + Mutate<Self::AccountId, Balance = Balance>

Assets for deposit/withdraw assets to/from pallet account

type RelayOrigin: EnsureOrigin<Self::Origin>

The origin which can do operation on relaychain using parachain’s sovereign account

type UpdateOrigin: EnsureOrigin<Self::Origin>

The origin which can update liquid currency, staking currency and other parameters

type PalletId: Get<PalletId>

The pallet id of liquid staking, keeps all the staking assets

type XcmSender: SendXcm

XCM message sender

type SelfParaId: Get<ParaId>

Returns the parachain ID we are running with.

type DerivativeIndex: Get<u16>

Account derivative index

type AccountIdToMultiLocation: Convert<Self::AccountId, MultiLocation>

Convert T::AccountId to MultiLocation.

type UnstakeQueueCapacity: Get<u32>

Unstake queue capacity

type MaxRewardsPerEra: Get<BalanceOf<Self>>

Max rewards per era

type MaxSlashesPerEra: Get<BalanceOf<Self>>

Max slashes per era

type MinStakeAmount: Get<BalanceOf<Self>>

Minimum stake amount

type MinUnstakeAmount: Get<BalanceOf<Self>>

Minimum unstake amount

type RelayNetwork: Get<NetworkId>

Relay network

type WeightInfo: WeightInfo

Weight information